[file] 인터랙션 모달 컴포넌트 분리
